撮影機能をViewModelからExtensionに移行
